test: changing assertion methods + refactoring
