+++ src/db/createConnection.ts
@@ -19,11 +19,13 @@ export const createConnectionOptions = (dbConfig: commonDbFullV1Type): PoolConfi
     password: dbConfig.password,
     database: dbConfig.database,
   };
+
   const sslParams = dbConfig.ssl;
   if (sslParams.enabled) {
     // todo - should be tested on future development (current version not support cert full deployment)
     dataSourceOptions.ssl = { key: readFileSync(sslParams.key), cert: readFileSync(sslParams.cert), ca: readFileSync(sslParams.ca) };
   }
+
   return dataSourceOptions;
 };
