+++ tests/unit/jobs/jobs.spec.ts
@@ -1,5 +1,6 @@
 import jsLogger from '@map-colonies/js-logger';
-import { Creator, JobMode, JobName, Prisma, PrismaClient } from '@prisma/client';
+import type { Creator, JobMode, JobName, Prisma } from '@prisma/client';
+import { PrismaClient } from '@prisma/client';
 
 import { JobManager } from '@src/jobs/models/jobManager';
 
@@ -37,68 +38,77 @@ describe('JobManager', () => {
   });
   describe('#Jobs', () => {
     describe('#createJob', () => {
-      it('should return created job formatted', async function () {
-        const jobEntity = createJobEntity({});
-        const prismaCreateJobMock = jest.spyOn(prisma.job, 'create').mockResolvedValue(jobEntity);
-        const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
+      describe('#HappyPath', () => {
+        it('should return created job formatted', async function () {
+          const jobEntity = createJobEntity({});
+          const prismaCreateJobMock = jest.spyOn(prisma.job, 'create').mockResolvedValue(jobEntity);
+          const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
 
-        const createJobParams = {
-          name: 'DEFAULT' as JobName,
-          creator: 'UNKNOWN' as Creator,
-          data: { stages: [] },
-          type: 'PRE_DEFINED' as JobMode,
-          notifications: {},
-          userMetadata: {},
-        };
+          const createJobParams = {
+            name: 'DEFAULT' as JobName,
+            creator: 'UNKNOWN' as Creator,
+            data: { stages: [] },
+            type: 'PRE_DEFINED' as JobMode,
+            notifications: {},
+            userMetadata: {},
+          };
 
-        const job = await jobManager.createJob(createJobParams);
+          const job = await jobManager.createJob(createJobParams);
 
-        expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
-        expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(1);
-        expect(job).toStrictEqual(jobManager.convertPrismaToJobResponse(jobEntity));
+          expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
+          expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(1);
+          expect(job).toStrictEqual(jobManager.convertPrismaToJobResponse(jobEntity));
+        });
       });
 
-      it('should failed on db error when creating job', async function () {
-        const prismaCreateJobMock = jest.spyOn(prisma.job, 'create').mockRejectedValueOnce(new Error('db connection error'));
-        const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
+      describe('#SadPath', () => {
+        it('should failed on db error when creating job', async function () {
+          const prismaCreateJobMock = jest.spyOn(prisma.job, 'create').mockRejectedValueOnce(new Error('db connection error'));
+          const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
 
-        const createJobParams = {
-          name: 'DEFAULT' as JobName,
-          creator: 'UNKNOWN' as Creator,
-          data: { stages: [] },
-          type: 'PRE_DEFINED' as JobMode,
-          notifications: {},
-          userMetadata: {},
-        };
+          const createJobParams = {
+            name: 'DEFAULT' as JobName,
+            creator: 'UNKNOWN' as Creator,
+            data: { stages: [] },
+            type: 'PRE_DEFINED' as JobMode,
+            notifications: {},
+            userMetadata: {},
+          };
 
-        await expect(jobManager.createJob(createJobParams)).rejects.toThrow('db connection error');
+          await expect(jobManager.createJob(createJobParams)).rejects.toThrow('db connection error');
 
-        expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
-        expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(0);
+          expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
+          expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(0);
+        });
       });
     });
+
     describe('#findJobs', () => {
-      it('should return array with single job formatted object by criteria', async function () {
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        const jobEntity = createJobEntity({ expirationTime: undefined, ttl: undefined });
-        const prismaCreateJobMock = jest.spyOn(prisma.job, 'findMany').mockResolvedValue([jobEntity]);
-        const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
+      describe('#HappyPath', () => {
+        it('should return array with single job formatted object by criteria', async function () {
+          // eslint-disable-next-line @typescript-eslint/naming-convention
+          const jobEntity = createJobEntity({ expirationTime: undefined, ttl: undefined });
+          const prismaCreateJobMock = jest.spyOn(prisma.job, 'findMany').mockResolvedValue([jobEntity]);
+          const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
 
-        const job = await jobManager.getJobs({ creator: 'UNKNOWN' });
+          const job = await jobManager.getJobs({ creator: 'UNKNOWN' });
 
-        expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
-        expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(1);
-        expect(job).toStrictEqual([jobManager.convertPrismaToJobResponse(jobEntity)]);
+          expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
+          expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(1);
+          expect(job).toStrictEqual([jobManager.convertPrismaToJobResponse(jobEntity)]);
+        });
       });
 
-      it('should failed on db error when find jobs', async function () {
-        const prismaCreateJobMock = jest.spyOn(prisma.job, 'findMany').mockRejectedValueOnce(new Error('db connection error'));
-        const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
+      describe('#SadPath', () => {
+        it('should failed on db error when find jobs', async function () {
+          const prismaCreateJobMock = jest.spyOn(prisma.job, 'findMany').mockRejectedValueOnce(new Error('db connection error'));
+          const convertPrismaToJobResponseStub = jest.spyOn(jobManager, 'convertPrismaToJobResponse');
 
-        await expect(jobManager.getJobs({ creator: 'UNKNOWN' })).rejects.toThrow('db connection error');
+          await expect(jobManager.getJobs({ creator: 'UNKNOWN' })).rejects.toThrow('db connection error');
 
-        expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
-        expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(0);
+          expect(prismaCreateJobMock).toHaveBeenCalledTimes(1);
+          expect(convertPrismaToJobResponseStub).toHaveBeenCalledTimes(0);
+        });
       });
     });
   });
